+++ modules/characters_encoding/no-src/include/characters_encoding.hpp
@@ -62,6 +62,12 @@ wstring_to_utf8(const wchar_t* str);
 NLSCHARACTERS_ENCODING_IMPEXP std::vector<std::string>
 wstring_to_utf8(const std::vector<std::wstring>& strs);
 //=============================================================================
+#ifdef _MSC_BUILD
+NLSCHARACTERS_ENCODING_IMPEXP
+bool
+utf8_to_wstring_Windows(const std::string& str, std::wstring& wstr);
+#endif
+//=============================================================================
 NLSCHARACTERS_ENCODING_IMPEXP bool
 isUnicodeLetter(wchar_t character);
 //=============================================================================